Enum jcli_lib::jcli_lib::certificate::NewArgs

pub enum NewArgs {
    StakePoolRegistration(StakePoolRegistration),
    StakeDelegation(StakeDelegation),
    OwnerStakeDelegation(OwnerStakeDelegation),
    StakePoolRetirement(StakePoolRetirement),
    VotePlan(VotePlanRegistration),
    VoteTally(VoteTallyRegistration),
    UpdateVote(UpdateVote),
    UpdateProposal(UpdateProposal),
    VoteCast(VoteCastCmd),
}

Variants

StakePoolRegistration(StakePoolRegistration)

create the stake pool registration certificate.

This contains all the declaration data of a stake pool. Including the management data. Once registered and accepted the users can delegate stake to the stake pool by referring the stake pool id.

--tax-* parameters allow to set the rewards the stake pool will take before serving the stake delegators. If the total reward for a stake pool is Y. The stake pool will take a fixed (--tax-fixed) first: X. Then will take a percentage of the remaining rewards (--tax-ratio): R. The total of the rewards gained from R can be capped by an optional --tax-limit: L where the actual tax T is X plus the minimum of L and R.

Delegators will then receive a share of the remaining rewards: Y - T.

StakeDelegation(StakeDelegation)

build a stake delegation certificate

OwnerStakeDelegation(OwnerStakeDelegation)

build an owner stake delegation certificate

StakePoolRetirement(StakePoolRetirement)

retire the given stake pool ID From the blockchain

by doing so all remaining stake delegated to this stake pool will become pending and will need to be re-delegated.

VotePlan(VotePlanRegistration)

create a new vote plan certificate

VoteTally(VoteTallyRegistration)

create a new vote tally certificate

UpdateVote(UpdateVote)

create a new update vote certificate

UpdateProposal(UpdateProposal)

create a new update proposal certificate

VoteCast(VoteCastCmd)

create a vote cast certificate

Implementations

impl NewArgs

pub fn exec(self) -> Result<(), Error>