Enum jcli_lib::jcli_lib::rest::v0::message::Message

pub enum Message {
    Post {
        args: RestArgs,
        file: Option<PathBuf>,
    },
    Logs {
        args: RestArgs,
        output_format: OutputFormat,
    },
}

Variants

Post

Fields

args: RestArgs

file: Option<PathBuf>

File containing hex-encoded message. If not provided, message will be read from stdin.

Post message. Prints id for posted message

Logs

Fields

args: RestArgs

output_format: OutputFormat

get the node’s logs on the message pool. This will provide information on pending transaction, rejected transaction and or when a transaction has been added in a block

Implementations

impl Message

pub fn exec(self) -> Result<(), Error>